etl/requirements.txt
pyodbc==5.1.0
sqlalchemy==2.0.25
python-dotenv==1.0.1
pandas==2.2.2
pyarrow>=21.0.0
sendgrid==6.11.0
sentry-sdk>=2.0.0
requests>=2.32.0  # Required for KLK API HTTP calls
psycopg2-binary>=2.9.0  # Required for PostgreSQL connection
